Add files via upload
The interface commands and the DAQ control commands are now in separated files.  I added some comments to the code and adjusted then according to PEP8. To the next step Taks and ChannelList will no longer be singletons allowing user to create more than one task and ChannelList will become part of the task properties. Output commands still need to be added and there are still some functionalities to be added